services/media_service.py
"""
媒体处理服务
"""

import logging
from typing import Optional
from telegram import Voice, PhotoSize, Bot

logger = logging.getLogger(__name__)


class MediaService:
    """媒体处理服务类"""

    def __init__(self):
        pass

    async def voice_to_text(self, voice: Voice, bot: Bot) -> Optional[str]:
        """语音转文字"""
        try:
            # 获取语音文件
            voice_file = await bot.get_file(voice.file_id)

            # 这里需要集成语音识别API
            # 例如：OpenAI Whisper, Google Speech-to-Text, Azure Speech Services

            # 示例实现（需要替换为真实的API调用）
            logger.info(f"处理语音文件: {voice_file.file_path}")

            # 模拟语音识别结果
            return "语音转文字功能需要集成相应的API服务"

        except Exception as e:
            logger.error(f"语音转文字处理失败: {e}")
            return None

    async def analyze_image(self, photo: PhotoSize, bot: Bot) -> Optional[str]:
        """分析图片内容"""
        try:
            # 获取图片文件
            photo_file = await bot.get_file(photo.file_id)

            # 这里需要集成图像识别API
            # 例如：OpenAI GPT-4 Vision, Google Vision API, Azure Computer Vision

            # 示例实现（需要替换为真实的API调用）
            logger.info(f"分析图片文件: {photo_file.file_path}")

            # 模拟图片分析结果
            return "图片分析功能需要集成相应的API服务。\n\n这是一张图片，包含了一些内容。如需详细分析，请配置图像识别API。"

        except Exception as e:
            logger.error(f"图片分析处理失败: {e}")
            return None

    async def process_document(self, document, bot: Bot) -> Optional[str]:
        """处理文档"""
        try:
            # 获取文档文件
            doc_file = await bot.get_file(document.file_id)

            # 根据文件类型处理
            file_name = document.file_name.lower()

            if file_name.endswith(('.txt', '.md')):
                return await self.process_text_document(doc_file)
            elif file_name.endswith(('.pdf')):
                return await self.process_pdf_document(doc_file)
            elif file_name.endswith(('.doc', '.docx')):
                return await self.process_word_document(doc_file)
            else:
                return f"暂不支持 {document.file_name} 类型的文档"

        except Exception as e:
            logger.error(f"文档处理失败: {e}")
            return None

    async def process_text_document(self, file) -> str:
        """处理文本文档"""
        # 实现文本文档处理逻辑
        return "文本文档处理功能开发中"

    async def process_pdf_document(self, file) -> str:
        """处理PDF文档"""
        # 实现PDF文档处理逻辑
        return "PDF文档处理功能开发中"

    async def process_word_document(self, file) -> str:
        """处理Word文档"""
        # 实现Word文档处理逻辑
        return "Word文档处理功能开发中"